* Patched modifications from OpenTibia repository * When namelocked player changes name, doors of his house will be now updated * Moved skulls to creature class what allow us to use it for monsters and npcs * Monster & NPC skulls support * Lua functions for skulls * Goods map isn't now filled with buyable items * Added some missing monsters, mostly of new Tibia (bosses and elementals)- could someone format them properly? thx
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@426 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ fileloader.cpp
@@ -557,7 +557,7 @@ int32_t FileLoader::loadCacheBlock(uint32_t pos)
 		return -1;
 	}
 
-	size_t size = fread(m_cached_data[loading_cache].data, 1, m_cache_size, m_file);
+	int32_t size = fread(m_cached_data[loading_cache].data, 1, m_cache_size, m_file);
 	m_cached_data[loading_cache].size = size;
 
 	if(size < (pos - m_cached_data[loading_cache].base))
